test(core): migrate TreeOwnsPanesTests (Wave B)
13 tests, 77 failure sites. Preserves the tree-owns-panes
structural-guarantee net: pane content lives in the leaf, lookups go
through model.pane(_:)/allPaneIds, and move/split/swap helpers thread
the full PaneModel payload (the splitRight regression test in particular).

Gate results:
- NAME parity:      13/13 baseline names found in migrated @Tests
- ASSERTION parity: baseline TOTAL 77 == new TOTAL 77
                    (63 #expect + 6 expectNoDifference + 3 #require
                    + 5 Issue.record). expectNoDifference replaces
                    expectEqual on PaneModel/AppModel value comparisons
                    so a failure surfaces a field-level diff (per plan
                    'apply expectNoDifference to big-tree assertions
                    here'). #require unwraps the validateAndBuild
                    optional; Issue.record + return preserves the
                    compound `guard case .split(...)` flow that
                    #require cannot pattern-match.
- STYLE:            13/13 tests carry Intent/Why/Scenario preambles

Verified locally:
- swift test --package-path lib/DanTermCore -> 73/73 in 8 suites (0.61s)
- ./test.sh -> 1005/1005 (= 1018 - 13)
